rename module to mime
