+++ src/js/defaults.js
@@ -44,7 +44,7 @@ const pluginDefaultOptions = {
     // The mime type for the video recorder.
     // Use 'video/mp4' (Firefox) or 'video/webm;codecs=H264' (Chrome 52 and
     // newer) for MP4.
-    videoMimeType: 'video/webm;codecs=vp8',
+    videoMimeType: 'video/webm',
     // Video recorder type to use. This allows you to specify an alternative
     // recorder class, e.g. WhammyRecorder from recordrtc. Defaults to 'auto'
     // which let's recordrtc specify the best available recorder type.
