+++ test/plugins/libvorbis-plugin.spec.js
@@ -2,13 +2,13 @@
  * @since 2.2.0
  */
 
-import TestHelpers from '../test-helpers.js';
+import TestHelpers from '../test-helpers';
 
-import Event from '../../src/js/event.js';
+import Event from '../../src/js/event';
 
 // registers the plugin
-import LibVorbisEngine from '../../src/js/plugins/libvorbis-plugin.js';
-import {LIBVORBISJS} from '../../src/js/engine/record-engine.js';
+import LibVorbisEngine from '../../src/js/plugins/libvorbis-plugin';
+import {LIBVORBISJS} from '../../src/js/engine/record-engine';
 
 
 /** @test {LibVorbisEngine} */
@@ -27,6 +27,7 @@ describe('plugins.libvorbis-plugin', () => {
         player.one(Event.FINISH_RECORD, () => {
             // received a blob file
             expect(player.recordedData instanceof Blob).toBeTruthy();
+            expect(player.recordedData.name).toEndWith('.oga');
 
             // wait till it's loaded before destroying
             // (XXX: create new event for this)
