+++ test/utils/browser-shim.spec.js
@@ -2,7 +2,7 @@
  * @since 2.2.0
  */
 
-import setSrcObject from '../../src/js/utils/browser-shim.js';
+import setSrcObject from '../../src/js/utils/browser-shim';
 
 /** @test {browser-shim} */
 describe('utils.setSrcObject', () => {
